Project/requirements.txt
Django~=3.2.7
psycopg2~=2.9.1
Pillow~=8.3.2
requests~=2.26.0
djangorestframework~=3.12.4
Markdown~=3.3.4